backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
email-validator==2.1.0